requirements_no_version.txt
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
arabic-reshaper
asgiref
asn1crypto
attrs
beautifulsoup4
blinker
Brotli
cachetools
certifi
cffi
chardet
charset-normalizer
click
colorama
crispy-bootstrap5
crispy-tailwind
cryptography
cssselect2
defusedxml
distlib
dj-database-url
Django
django-allauth
django-bootstrap-v5
django-bootstrap5
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
django-rest-auth
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
EasyProcess
entrypoint2
filelock
Flask
fonttools
frozenlist
git-filter-repo
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
gunicorn
html5lib
httplib2
idna
itsdangerous
Jinja2
lxml
Markdown
MarkupSafe
MouseInfo
mss
multidict
mysql-connector-python
numpy
oauth2client
oauthlib
opencv-python
oscrypto
packaging
phonenumbers
Pillow
pipenv
platformdirs
proto-plus
protobuf
psycopg2
psycopg2-binary
pyasn1
pyasn1_modules
PyAutoGUI
pycparser
PyDrive
pydyf
PyGetWindow
pyHanko
pyhanko-certvalidator
PyJWT
PyMsgBox
PyMySQL
pyparsing
pypdf
pyperclip
pyphen
pypng
PyRect
pyscreenshot
PyScreeze
pytesseract
python-bidi
python-decouple
python3-openid
pytweening
pywin32
PyYAML
qrcode
reportlab
requests
requests-oauthlib
rsa
six
social-auth-app-django
social-auth-core
soupsieve
sqlparse
sslcommerz-lib
svglib
tinycss2
twilio
typing_extensions
tzdata
tzlocal
uritemplate
uritools
urllib3
virtualenv
webencodings
Werkzeug
whitenoise
xhtml2pdf
yarl
zopfli
